Moved AddSpecialist, removed country at AddSpecialist
