+++ src/components/Admin/AdminDashboard.js
@@ -4,6 +4,7 @@ import { baseUrl } from "../../constants";
 import { Redirect, Link } from "react-router-dom";
 import "../MyIdea/Dashboard/IdeaDashboard.css";
 import posed from "react-pose";
+import mentor from "../../res/mentor.png";
 
 export default function AdminDashboard(props) {
   const [user, setUserData] = useState({});
@@ -31,6 +32,20 @@ export default function AdminDashboard(props) {
     props.user();
   }
 
+  const showNewSpecialist = () => {
+    if (!props.authState.user) return null;
+    if (props.authState.user.role === "admin") {
+      return (
+        <Link className="links" to="/Admin/dashboard/newspecialist">
+          <div className="invest-tile">
+            <img className="icons" src={mentor}></img>
+            <h4>Add Specialist</h4>
+          </div>
+        </Link>
+      );
+    } else return null;
+  };
+
   return (
     <div className="dashboard-container">
       <br />
@@ -39,6 +54,7 @@ export default function AdminDashboard(props) {
       <div className="title">
         <h1>{user.firstName}'s Admin Dashboard</h1>
       </div>
+      <div className="flex-tilescontainer">{showNewSpecialist()}</div>
       <h2 style={styledH2}>Ideas:</h2>
 
       <div className="flex-tilescontainer">
